# Speed up loading blocks using init with meta weights