@borzunov 2023-03-12
Awesome results!

The total block loading time (including the time to move it on GPU) went down from 57 sec to 25 sec on my machines. This means that, given that the blocks are already downloaded, the servers will spend 2x less time for restarting after being preempted or during rebalancing.